Netflix Data Analysis
The objective of this project is to examine the content released by Netflix and derive insights that can guide the company's future direction. Historically, the streaming service has primarily emphasized movie content, but upon taking a comprehensive look at the data, it becomes evident that the realm of TV shows and series has experienced significant growth.